requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==22.2.0
    # via pytest
audioread==3.0.0
    # via librosa
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via soundfile
charset-normalizer==3.1.0
    # via requests
decorator==5.1.1
    # via librosa
exceptiongroup==1.1.0
    # via pytest
ffmpeg-python==0.2.0
    # via -r requirements.in
future==0.18.3
    # via ffmpeg-python
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
joblib==1.2.0
    # via
    #   librosa
    #   scikit-learn
lazy-loader==0.1
    # via librosa
librosa==0.10.0
    # via -r requirements.in
linkify-it-py==2.0.0
    # via markdown-it-py
llvmlite==0.39.1
    # via numba
lxml==4.9.2
    # via -r requirements.in
markdown-it-py[linkify]==2.2.0
    # via
    #   -r requirements.in
    #   mdformat
    #   mdformat-gfm
    #   mdit-py-plugins
mdformat==0.7.16
    # via
    #   -r requirements.in
    #   mdformat-gfm
    #   mdformat-tables
mdformat-gfm==0.3.5
    # via -r requirements.in
mdformat-tables==0.4.1
    # via mdformat-gfm
mdit-py-plugins==0.3.5
    # via mdformat-gfm
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.0.5
    # via librosa
numba==0.56.4
    # via librosa
numpy==1.23.5
    # via
    #   -r requirements.in
    #   librosa
    #   numba
    #   opencv-python
    #   scikit-learn
    #   scipy
    #   soxr
opencv-python==4.7.0.72
    # via -r requirements.in
packaging==23.0
    # via
    #   pooch
    #   pytest
pillow==9.4.0
    # via -r requirements.in
platformdirs==3.1.1
    # via pooch
pluggy==1.0.0
    # via pytest
pooch==1.7.0
    # via librosa
pycparser==2.21
    # via cffi
pytest==7.2.2
    # via -r requirements.in
python-magic==0.4.27
    # via -r requirements.in
requests==2.28.2
    # via pooch
scikit-learn==1.2.2
    # via librosa
scipy==1.10.1
    # via
    #   librosa
    #   scikit-learn
soundfile==0.12.1
    # via librosa
soxr==0.3.4
    # via librosa
threadpoolctl==3.1.0
    # via scikit-learn
tomli==2.0.1
    # via
    #   mdformat
    #   pytest
typing-extensions==4.5.0
    # via librosa
uc-micro-py==1.0.1
    # via linkify-it-py
urllib3==1.26.15
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
